Read stories and requirements from YAML files

Convert a set of stories and optionally requirements from YAML files into a data frame that is ready to be passed as the specs argument to mrgvalidate::create_validation_docs().

read_spec_yaml(stories, requirements = NULL)

Arguments

stories, requirements

Character vector of YAML files to read stories and requirements from.

Value

Tibble with columns as described by mrgvalidate::input_formats.

Details

Option 1: Stories and requirements

If you provide a value for both stories and requirements, the stories argument should be one or more YAML files with the following structure:

The requirements argument should be a set of YAML files that have entries that link those requirements to tests:

The information above is used to construct a data frame with the following columns:

StoryId (character scalar)

StoryName (character scalar)

StoryDescription (character scalar)

ProductRisk (character scalar)

RequirementIds (character vector)

RequirementDescription (character scalar)

TestIds (character vector)

Option 2: Stories linked directly to tests

If you provide only a stories argument, the YAML entries should specify a list of tests rather than a set of requirements:

Those entries are used to construct a data frame with the following columns:

StoryId (character scalar)

StoryName (character scalar)

StoryDescription (character scalar)

ProductRisk (character scalar)

TestIds (character vector)